backend/requirements.txt
asgiref==3.11.0
Django==5.2.12
django-mongodb-backend==5.2.3
pymongo==4.16.0
dnspython==2.8.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.0.3
packaging==26.0
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
certifi==2025.1.31

PyJWT==2.8.0
cryptography==41.0.0
requests==2.31.0
resend==2.4.0
django-ratelimit==4.1.0
